+++ apps/dotcom/client/src/tla/components/dialogs/SubmitFeedbackDialog.tsx
@@ -50,7 +50,7 @@ function SignedOutSubmitFeedbackDialog({ onClose }: { onClose(): void }) {
 				</p>
 				<ul style={{ gap: 4 }}>
 					<li>
-						<ExternalLink to="https://discord.gg/rhsyWMUJxd">
+						<ExternalLink to="https://discord.tldraw.com/?utm_source=dotcom&utm_medium=organic&utm_campaign=dotcom-feedback">
 							<F defaultMessage="Chat with us on Discord" />
 						</ExternalLink>
 					</li>
@@ -125,7 +125,11 @@ function SignedInSubmitFeedbackDialog({ onClose }: { onClose(): void }) {
 						defaultMessage="Fill out this form, <discord>chat with us on Discord</discord> or <github>submit an issue on GitHub</github>."
 						values={{
 							discord: (chunks) => {
-								return <ExternalLink to="https://discord.gg/rhsyWMUJxd">{chunks}</ExternalLink>
+								return (
+									<ExternalLink to="https://discord.tldraw.com/?utm_source=dotcom&utm_medium=organic&utm_campaign=dotcom-feedback">
+										{chunks}
+									</ExternalLink>
+								)
 							},
 							github: (chunks) => {
 								return (